refactor: http 패키지 생성
- HTTP Request와 HTTP Response와 관련된 클래스를 http 패키지로 이동
