+++ src/main/java/common/http/ResponseLine.java
@@ -1,4 +1,4 @@
-package common;
+package common.http;
 
 import common.enums.ResponseCode;
