+++ src/main/java/controller/IndexController.java
@@ -1,7 +1,7 @@
 package controller;
 
 import modelview.ModelView;
-import common.HttpRequest;
+import common.http.HttpRequest;
 
 import static common.enums.Method.isGetMethod;
