+++ src/main/java/controller/StaticController.java
@@ -1,7 +1,7 @@
 package controller;
 
 import modelview.ModelView;
-import common.HttpRequest;
+import common.http.HttpRequest;
 
 public class StaticController implements Controller {
     @Override
