+++ src/main/java/webserver/ViewResolver.java
@@ -8,6 +8,9 @@ import static webserver.ServerConfig.TEMPLATE_PATH;
 
 public class ViewResolver {
     public static View resolve(ContentType contentType, String viewName) {
+        if (viewName == null) {
+            return new View(null);
+        }
         if (contentType.isStaticContent()) {
             return new View(STATIC_PATH + viewName);
         }
